Teleselection

+++ Assets/_game/Mechanics/Aether/Components/CSpawning.cs
@@ -3,7 +3,7 @@
 public class CSpawning : MonoBehaviour
 {
     public GameObject prefabToSpawn;
-    public int spawnTimer;
+    public float spawnTimer;
     public float timeOfLastSpawn;
     public float forwardVelocity;
 }
